fix: flash mint maximum is uint112.max

+++ contracts/WETH10.sol
@@ -103,7 +103,7 @@ contract WETH10 is IWETH10 {
     /// The flash minted WETH10 is not backed by real Ether, but can be withdrawn as such up to the Ether balance of this contract.
     /// Arbitrary data can be passed as a bytes calldata parameter.
     /// Emits two {Transfer} events for minting and burning of the flash minted amount.
-    function flashMint(uint256 value, bytes calldata data) external override {
+    function flashMint(uint112 value, bytes calldata data) external override {
         flashSupply += value;
         require(address(this).balance + flashSupply <= type(uint112).max, "WETH::flashMint: supply limit exceeded");
         balanceOf[msg.sender] += value;
