+++ contracts/tests/TestFlashMinter.sol
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity 0.7.0;
 
+
 interface FlashMintableLike {
-    function flashMint(uint256, bytes calldata) external;
+    function flashMint(uint112, bytes calldata) external;
     function balanceOf(address) external returns (uint256);
     function deposit() external payable;
     function withdraw(uint256) external;
@@ -19,7 +20,7 @@ contract TestFlashMinter {
     receive() external payable {}
 
     function executeOnFlashMint(bytes calldata data) external {
-        (Action action, address user, uint256 value) = abi.decode(data, (Action, address, uint256)); // Use this to unpack arbitrary data
+        (Action action, address user, uint112 value) = abi.decode(data, (Action, address, uint112)); // Use this to unpack arbitrary data
         flashUser = user;
         flashValue = value;
         if (action == Action.NORMAL) {
@@ -37,31 +38,31 @@ contract TestFlashMinter {
         }
     }
 
-    function flashMint(address target, uint256 value) public {
+    function flashMint(address target, uint112 value) public {
         // Use this to pack arbitrary data to `executeOnFlashMint`
         bytes memory data = abi.encode(Action.NORMAL, msg.sender, value); // Here msg.sender is the user, and target is the weth contract
         FlashMintableLike(target).flashMint(value, data);
     }
 
-    function flashMintAndWithdraw(address target, uint256 value) public {
+    function flashMintAndWithdraw(address target, uint112 value) public {
         // Use this to pack arbitrary data to `executeOnFlashMint`
         bytes memory data = abi.encode(Action.WITHDRAW, msg.sender, value); // Here msg.sender is the user, and target is the weth contract
         FlashMintableLike(target).flashMint(value, data);
     }
 
-    function flashMintAndSteal(address target, uint256 value) public {
+    function flashMintAndSteal(address target, uint112 value) public {
         // Use this to pack arbitrary data to `executeOnFlashMint`
         bytes memory data = abi.encode(Action.STEAL, msg.sender, value); // Here msg.sender is the user, and target is the weth contract
         FlashMintableLike(target).flashMint(value, data);
     }
 
-    function flashMintAndReenter(address target, uint256 value) public {
+    function flashMintAndReenter(address target, uint112 value) public {
         // Use this to pack arbitrary data to `executeOnFlashMint`
         bytes memory data = abi.encode(Action.REENTER, msg.sender, value); // Here msg.sender is the user, and target is the weth contract
         FlashMintableLike(target).flashMint(value, data);
     }
 
-    function flashMintAndOverspend(address target, uint256 value) public {
+    function flashMintAndOverspend(address target, uint112 value) public {
         bytes memory data = abi.encode(Action.OVERSPEND, msg.sender, value); // Here msg.sender is the user, and target is the weth contract
         FlashMintableLike(target).flashMint(value, data);
     }
